refactor(cli): move implementation into src module
